[#127] reorganized route testing
